Changes in response to CR
