+++ src/main/java/com/gmail/goosius/siegewar/timeractions/DefenderWin.java
@@ -5,7 +5,7 @@ import com.gmail.goosius.siegewar.enums.SiegeStatus;
 import com.gmail.goosius.siegewar.objects.Siege;
 import com.gmail.goosius.siegewar.utils.SiegeWarMoneyUtil;
 import com.gmail.goosius.siegewar.utils.SiegeWarSiegeCompletionUtil;
-import com.gmail.goosius.siegewar.utils.SiegeWarTimeUtil;
+import com.gmail.goosius.siegewar.utils.SiegeWarImmunityUtil;
 
 /**
  * This class is responsible for processing siege defender wins
@@ -34,7 +34,7 @@ public class DefenderWin
 				TownOccupationController.removeTownOccupation(siege.getTown());
 				break;
 			case REVOLT:
-				SiegeWarTimeUtil.activateRevoltImmunityTimer(siege.getTown());
+				SiegeWarImmunityUtil.activateRevoltImmunityTimer(siege.getTown());
 				break;
 		}
     }
